requirements.txt
django-environ==0.4.5
django==3.2.13
djangorestframework==3.12.2
psycopg2-binary==2.8.6
django-filter==21.1
djangorestframework-api-key==2.2.0
